GUI-2282 reworked test for Alice
